Fix type precision in dorbdb6 and zunbdb6

+++ SRC/dorbdb6.f
@@ -181,7 +181,7 @@
 *     ..
 *     .. Local Scalars ..
       INTEGER            I, IX
-      REAL               EPS, NORM, NORM_NEW, SCL, SSQ
+      DOUBLE PRECISION   EPS, NORM, NORM_NEW, SCL, SSQ
 *     ..
 *     .. External Functions ..
       DOUBLE PRECISION   DLAMCH
